+++ packages/components/src/components/callout/callout.spec.ts
@@ -1,32 +1,32 @@
 import { newSpecPage } from '@stencil/core/testing';
-import { Badge } from './badge';
+import { Callout } from './callout';
 
 it('should reflect attributes/props', async () => {
   const page = await newSpecPage({
-    components: [Badge],
-    html: `<scale-badge
+    components: [Callout],
+    html: `<scale-callout
                 size ="large"
-                color ="primary"
+                variant ="primary"
                 rotation ="15">
-              </scale-badge>`,
+              </scale-callout>`,
   });
 
   expect(page.rootInstance.size).toBe('large');
-  expect(page.rootInstance.color).toBe('primary');
+  expect(page.rootInstance.variant).toBe('primary');
   expect(page.rootInstance.rotation).toBe(15);
 });
 
-it('checks another color, other than prop', async () => {
+it('checks another variant, other than prop', async () => {
   const page = await newSpecPage({
-    components: [Badge],
-    html: `<scale-badge
+    components: [Callout],
+    html: `<scale-callout
                 size ="large"
-                color ="red"
+                variant ="red"
                 rotation ="15">
-              </scale-badge>`,
+              </scale-callout>`,
   });
 
   expect(page.rootInstance.size).toBe('large');
-  expect(page.rootInstance.color).toBe('red');
+  expect(page.rootInstance.variant).toBe('red');
   expect(page.rootInstance.rotation).toBe(15);
 });